Reorganized Operator block test cases + new tests for DTO factories
