+++ tests/Operator/BranchDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Operator;
+
+use Grixu\SociusModels\Operator\Factories\BranchDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class BranchDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = BranchDataFactory::class;
+}
