+++ tests/Operator/BranchMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Operator;
 
 use CreateBranchesTable;
 use Grixu\SociusModels\Tests\Helpers\MigrationTest;
@@ -13,7 +13,7 @@ class BranchMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/operator/2020_09_30_135749_create_branches_table.php';
+        require_once __DIR__ . '/../../migrations/operator/2020_09_30_135749_create_branches_table.php';
         $this->obj = new CreateBranchesTable();
     }
 }
