+++ tests/Operator/OperatorBranchMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Operator;
 
 use CreateOperatorBranchPivotTable;
 use Grixu\SociusModels\Tests\Helpers\MigrationTest;
@@ -14,7 +14,7 @@ class OperatorBranchMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/operator/2020_10_01_064502_create_operator_branch_pivot_table.php';
+        require_once __DIR__ . '/../../migrations/operator/2020_10_01_064502_create_operator_branch_pivot_table.php';
         $this->obj = new CreateOperatorBranchPivotTable();
     }
 }
