+++ tests/Operator/OperatorDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Operator;
+
+use Grixu\SociusModels\Operator\Factories\OperatorDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class OperatorDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = OperatorDataFactory::class;
+}
