+++ tests/Operator/OperatorMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Operator;
 
 use CreateOperatorRolesTable;
 use CreateOperatorsTable;
@@ -15,14 +15,14 @@ class OperatorMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/operator/2020_09_30_092119_create_operators_table.php';
+        require_once __DIR__ . '/../../migrations/operator/2020_09_30_092119_create_operators_table.php';
         $this->obj = new CreateOperatorsTable();
     }
 
     /** @test */
     public function role_relationship(): void
     {
-        require_once __DIR__ . '/../migrations/operator/2020_09_30_082556_create_operator_roles_table.php';
+        require_once __DIR__ . '/../../migrations/operator/2020_09_30_082556_create_operator_roles_table.php';
         (new CreateOperatorRolesTable())->up();
 
         $this->up();
