+++ tests/Operator/OperatorRoleDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Operator;
+
+use Grixu\SociusModels\Operator\Factories\OperatorRoleDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class OperatorRoleDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = OperatorRoleDataFactory::class;
+}
