+++ tests/Operator/OperatorRoleMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Operator;
 
 use CreateOperatorRolesTable;
 use Grixu\SociusModels\Tests\Helpers\MigrationTest;
@@ -13,7 +13,7 @@ class OperatorRoleMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/operator/2020_09_30_082556_create_operator_roles_table.php';
+        require_once __DIR__ . '/../../migrations/operator/2020_09_30_082556_create_operator_roles_table.php';
         $this->obj = new CreateOperatorRolesTable();
     }
 }
